refactor: extract workspace filter into a separate component
